navigation graph
- was added navigation graph to navigation bottom bar
