+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:name=".App"
         tools:targetApi="31">
         <activity
-            android:name=".main.MainActivity"
+            android:name=".ui.MainActivity"
 
             android:exported="true">
             <intent-filter>
