--- app/src/main/java/com/example/businesstaxcalculator/main/MainActivity.kt
@@ -1,64 +1,0 @@
-package com.example.businesstaxcalculator.main
-
-import android.os.Bundle
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
-import androidx.fragment.app.Fragment
-import com.example.businesstaxcalculator.R
-import com.example.businesstaxcalculator.databinding.ActivityMainBinding
-import com.example.businesstaxcalculator.main.home.HomeFragment
-import com.example.businesstaxcalculator.main.profile.ProfileFragment
-import com.example.businesstaxcalculator.main.settings.SettingsFragment
-import com.google.android.material.bottomnavigation.BottomNavigationView
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    lateinit var bottomNavigationView: BottomNavigationView
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        enableEdgeToEdge()
-        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
-            val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
-            v.setPadding(systemBars.left, systemBars.top, systemBars.right, 0)
-            insets
-        }
-
-        loadFragment(HomeFragment())
-
-        binding.bottomNavigation.setOnItemSelectedListener {
-            when (it.itemId) {
-                R.id.home -> {
-                    loadFragment(HomeFragment())
-                    true
-                }
-                R.id.settings -> {
-                    loadFragment(SettingsFragment())
-                    true
-                }
-                R.id.profile -> {
-                    loadFragment(ProfileFragment())
-                    true
-                }
-                else -> error(R.string.fragment_error)
-            }
-        }
-    }
-
-    private fun loadFragment(fragment: Fragment) {
-
-        if (fragment != null) {
-            val transaction = supportFragmentManager.beginTransaction()
-            transaction.replace(R.id.container, fragment)
-            transaction.commit()
-        }
-    }
-}
