+++ app/src/main/java/com/example/businesstaxcalculator/ui/profile/ProfileFragment.kt
@@ -1,20 +1,20 @@
-package com.example.businesstaxcalculator.main.settings
+package com.example.businesstaxcalculator.ui.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.example.businesstaxcalculator.R
-import com.example.businesstaxcalculator.main.base.BaseTabFragment
+import com.example.businesstaxcalculator.ui.base.BaseTabFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class SettingsFragment : BaseTabFragment() {
+class ProfileFragment : BaseTabFragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        val view = inflater.inflate(R.layout.fragment_settings, container, false)
+        val view = inflater.inflate(R.layout.fragment_profile, container, false)
         return view
     }
 }
