+++ app/src/main/res/layout/home_fragment_layout.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools="com.example.businesstaxcalculator.main.home.HomeFragment">
+    tools="com.example.businesstaxcalculator.ui.home.HomeFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/fragment_container_view"
