+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/home">
+
+    <fragment
+        android:id="@+id/home"
+        android:name="com.example.businesstaxcalculator.ui.home.HomeFragment"
+        android:label="@string/home">
+
+        <action
+            android:id="@+id/action_to_home"
+            app:destination="@id/home" />
+
+        <action
+            android:id="@+id/action_to_profile"
+            app:destination="@id/profile" />
+
+        <action
+            android:id="@+id/action_to_settings"
+            app:destination="@id/settings" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/profile"
+        android:name="com.example.businesstaxcalculator.ui.profile.ProfileFragment"
+        android:label="@string/profile">
+        <action
+            android:id="@+id/action_to_home"
+            app:destination="@id/home" />
+
+        <action
+            android:id="@+id/action_to_profile"
+            app:destination="@id/profile" />
+
+        <action
+            android:id="@+id/action_to_settings"
+            app:destination="@id/settings" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/settings"
+        android:name="com.example.businesstaxcalculator.ui.settings.SettingsFragment"
+        android:label="@string/settings">
+        <action
+            android:id="@+id/action_to_home"
+            app:destination="@id/home" />
+
+        <action
+            android:id="@+id/action_to_profile"
+            app:destination="@id/profile" />
+
+        <action
+            android:id="@+id/action_to_settings"
+            app:destination="@id/settings" />
+    </fragment>
+
+
+</navigation>
